+++ src/main/java/com/etelligens/ecommerce/repositories/ProductImagesRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.etelligens.ecommerce.model.Images;
 
-public interface ProductImagesRepo extends JpaRepository<Images, Integer>{
+public interface ProductImagesRepo extends JpaRepository<Images, Long>{
 
 }
