+++ src/main/java/com/etelligens/ecommerce/repositories/ProductRepo.java
@@ -11,14 +11,10 @@ import org.springframework.stereotype.Repository;
 import com.etelligens.ecommerce.model.Product;
 
 @Repository
-public interface ProductRepo extends JpaRepository<Product, Integer> {
+public interface ProductRepo extends JpaRepository<Product, Long> {
 
 	public Optional<Product> findById(Integer productId);
-	
-//	@Query(value = "SELECT * FROM product WHERE category_id=:id  AND slug=:slugId", nativeQuery=true)
-//	public List<Product> findByCategoryOrSlug(Long id, Long slugId);
-	
-	
+
 	@Query(value = "SELECT * FROM product WHERE category_id=:id", nativeQuery=true)
 	public List<Product> findAllByCategoryId(Long id);
